feat: 한줄평 반응 수정 API 구현

+++ server/src/main/java/com/codecozy/server/dto/request/ReactionCommentRequest.java
@@ -2,5 +2,5 @@ package com.codecozy.server.dto.request;
 
 public record ReactionCommentRequest(
     String name,
-    long commentReaction
+    int commentReaction
 ) {}
